timer with new push ui restart
