+++ ProjectSettings/ProjectSettings.asset
@@ -121,8 +121,7 @@ PlayerSettings:
     16:9: 1
     Others: 1
   bundleVersion: 0.1
-  preloadedAssets:
-  - {fileID: -6602195166426926598, guid: 2ba4dbd235ed941bb8fdb5a40ab00f22, type: 2}
+  preloadedAssets: []
   metroInputSource: 0
   wsaTransparentSwapchain: 0
   m_HolographicPauseOnTrackingLoss: 1
@@ -644,7 +643,7 @@ PlayerSettings:
     m_ModelFolderPath: 
     m_PortalFolderPath: 
   luminCert:
-    m_CertPath: /Users/zacharyjohnson/Downloads/Coding Things/CS98/cookAlong/cook_along/privatekey/cookalongzdj.cert
+    m_CertPath: /Users/dani/Documents/College/19F/CS98/cubetemplate.cert
     m_SignPackage: 1
   luminIsChannelApp: 0
   luminVersion:
